+++ Fiap2NettTC3.sln
@@ -65,8 +65,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Gateway", "Gateway", "{3D30
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TicketNow.Gateway.Api", "Gateway\TicketNow.Gateway\TicketNow.Gateway.Api.csproj", "{4BB49041-4BB8-4CFC-884C-6D068863D810}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Payments.Api", "Payment\Payments.Api\Payments.Api.csproj", "{D861D939-7153-452D-9F1D-176395C3FBAC}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -173,10 +171,6 @@ Global
 		{4BB49041-4BB8-4CFC-884C-6D068863D810}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4BB49041-4BB8-4CFC-884C-6D068863D810}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4BB49041-4BB8-4CFC-884C-6D068863D810}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D861D939-7153-452D-9F1D-176395C3FBAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D861D939-7153-452D-9F1D-176395C3FBAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D861D939-7153-452D-9F1D-176395C3FBAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D861D939-7153-452D-9F1D-176395C3FBAC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -207,7 +201,6 @@ Global
 		{A31BE1C3-09A7-4B3D-B75B-AC949822E01B} = {19FA6AE1-0B1A-429A-A365-78099B8F3599}
 		{9FE4DA30-773F-478B-A18E-FE6FC77E2F5F} = {19FA6AE1-0B1A-429A-A365-78099B8F3599}
 		{4BB49041-4BB8-4CFC-884C-6D068863D810} = {3D3096C3-099E-4987-8494-638D0B37CAE2}
-		{D861D939-7153-452D-9F1D-176395C3FBAC} = {FD0B260F-9BFE-40A4-A65B-30B101E804ED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D3A9DA40-572B-4508-833C-CB9593B63D58}
