+++ Order/Order.Domain/Dtos/MockPayment/PaymentsDto.cs
@@ -1,13 +1,10 @@
 ﻿using Order.Domain.Enums;
 
-namespace Order.Domain.Dtos.MockPayment
+namespace Order.Domain.Dtos.MockPayment;
+
+public class PaymentsDto
 {
-	public class PaymentsDto
-	{
-		public int OrderId { get; set; }
-		public PaymentMethodEnum PaymentMethod { get; set; }
-		public PaymentStatusEnum PaymentStatus { get; set; }
-		public DateTime CreatedAt { get; set; }
-		public DateTime UpdatedAt { get; set;}
-	}
+	public int OrderId { get; set; }
+	public PaymentMethodEnum PaymentMethod { get; set; }
+	public PaymentStatusEnum PaymentStatus { get; set; }
 }
