+++ Payment/Payment.Api/appsettings.Development.json
@@ -8,8 +8,7 @@
   "AllowedHosts": "*",
 
   "ConnectionStrings": {
-    //"local": "Server=localhost;Database=ticket_now_order;User ID=sa;Password=1q2w3e4r@#$;Trusted_Connection=false;TrustServerCertificate=true;", //add to azure key vault,
-    "local": "Server=localhost;Database=ticket_now;Trusted_Connection=True;TrustServerCertificate=True"
+    "local": "Server=localhost;Database=ticket_now_order;User ID=sa;Password=1q2w3e4r@#$;Trusted_Connection=false;TrustServerCertificate=true;" //add to azure key vault,
   },
 
   "PaymentApi": {
