+++ Payment/Payment.Api/appsettings.json
@@ -9,7 +9,7 @@
   "AllowedHosts": "*",
 
   "ConnectionStrings": {
-    "local": "Server=localhost;Database=ticket_now_order;User ID=sa;Password=1q2w3e4r@#$;Trusted_Connection=false;TrustServerCertificate=true;", //add to azure key vault,
+    "local": "Server=localhost;Database=ticket_now_order;User ID=sa;Password=1q2w3e4r@#$;Trusted_Connection=false;TrustServerCertificate=true;" //add to azure key vault,
   },
 
   "PaymentApi": {
