+++ Payment/Payment.Domain/DTOs/PaymentsDto.cs
@@ -1,8 +1,8 @@
-﻿public class PaymentsDto
+﻿namespace Payment.Domain.DTOs;
+
+public class PaymentsDto
 {
     public int OrderId { get; set; }
     public PaymentMethod PaymentMethod { get; set; }
     public PaymentStatus PaymentStatus { get; set; }
-    public DateTime CreatedAt { get; set; }
-    public DateTime UpdatedAt { get; set; }
 }
